# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Meemo API
  version: 2.0.0
  description: >
    Meemo is a personal note and knowledge management service.
    This specification describes the authoritative HTTP REST API for Meemo v2.
servers:
  - url: /
    description: Current host server
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: connect.sid
      description: Session cookie generated after successful login.
  schemas:
    ErrorResponse:
      type: object
      required:
        - status
        - code
        - message
      properties:
        status:
          type: string
          example: Bad Request
        code:
          type: string
          example: invalid_request
          enum:
            - invalid_request
            - authentication_required
            - invalid_credentials
            - forbidden
            - not_found
            - conflict
            - payload_too_large
            - too_many_requests
            - internal_error
            - service_unavailable
            - request_failed
        message:
          type: string
          example: content must be a string
...
    PublicUserProfile:
      type: object
      additionalProperties: false
      required:
        - id
        - username
        - displayName
      properties:
        id:
          type: string
          example: 64f8a12b3c4d5e6f7a8b9c0d
        username:
          type: string
          example: alice
        displayName:
          type: string
          example: Alice Walker
    PublicUserSummary:
      type: object
      additionalProperties: false
      required:
        - username
        - displayName
      properties:
        username:
          type: string
          example: alice
        displayName:
          type: string
          example: Alice Walker
    AttachmentDescriptor:
      type: object
      required:
        - identifier
      properties:
        identifier:
          type: string
          description: Opaque storage UUID
          example: 9b1deb4d-3b7d-4bad-9bdd-2b0d7b3dcb6d
        fileName:
          type: string
          example: screenshot.png
        type:
          type: string
          enum: [image, unknown]
          example: image
        mime:
          type: string
          example: image/png
        size:
          type: integer
          example: 1048576
    ExternalContentItem:
      type: object
      required:
        - url
        - type
      properties:
        url:
          type: string
          format: uri
        type:
          type: string
          enum: [image, unknown]
    NoteColor:
      type: string
      enum: [default, coral, peach, sand, mint, sage, fog, storm, dusk, blossom, clay, chalk]
      example: default
    Thing:
      type: object
      required:
        - _id
        - ownerId
        - content
        - richContent
        - createdAt
        - modifiedAt
        - tags
        - attachments
        - public
        - shared
        - archived
        - sticky
      properties:
        _id:
          type: string
          example: 64f8a12b3c4d5e6f7a8b9c0d
        ownerId:
          type: string
          example: 64f8a12b3c4d5e6f7a8b9c01
        content:
          type: string
          example: Hello #meeting with @bob
        richContent:
          type: string
          example: Hello <a href="/tags/meeting">#meeting</a> with @bob
        createdAt:
          type: integer
          description: Unix timestamp in milliseconds
          example: 1694000000000
        modifiedAt:
          type: integer
          description: Unix timestamp in milliseconds of the latest complete note update
          example: 1694000000000
        tags:
          type: array
          items:
            type: string
          example: [meeting]
        externalContent:
          type: array
          items:
            $ref: '#/components/schemas/ExternalContentItem'
          default: []
        attachments:
          type: array
          items:
            $ref: '#/components/schemas/AttachmentDescriptor'
        public:
          type: boolean
          example: false
        shared:
          type: boolean
          example: false
        archived:
          type: boolean
          example: false
        sticky:
          type: boolean
          example: false
        color:
          allOf:
            - $ref: '#/components/schemas/NoteColor'
          default: default
        score:
          type: number
          description: Relevance match score when search mode is text
          example: 1.25
...
  parameters:
...
    PublicThingIdPath:
      name: thingId
      in: path
      required: true
      description: 24-character hexadecimal ObjectId of the Thing
      schema:
        type: string
        pattern: '^[0-9a-fA-F]{24}$'
        example: 64f8a12b3c4d5e6f7a8b9c0d
    UserIdPath:
      name: userId
      in: path
      required: true
      description: User identifier (stable ObjectId hex string or username)
      schema:
        type: string
        example: alice
...
    FilterQuery:
      name: filter
      in: query
      required: false
      description: Search query string supporting text keywords, #tags, and [[WikiLinks]]
      schema:
        type: string
        maxLength: 1000
    SearchModeQuery:
      name: mode
      in: query
      required: false
      description: Search execution mode ('regex' for substring pattern matching, 'text' for MongoDB full-text index with stemming and scoring)
      schema:
        type: string
        enum: [regex, text]
        default: regex
...
    SkipQuery:
      name: skip
      in: query
      required: false
      description: Number of records to skip
      schema:
        type: integer
        minimum: 0
        default: 0
    LimitQuery:
      name: limit
      in: query
      required: false
      description: Maximum records to return
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
  responses:
    BadRequest:
      description: Validation or parameter format error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    Forbidden:
      description: Access forbidden by authorization rules
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Requested resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalError:
      description: Unexpected internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
paths:
...
  /api/public/{userId}/things:
    get:
      tags: [public]
      summary: List public notes of a user
      parameters:
        - $ref: '#/components/parameters/UserIdPath'
        - $ref: '#/components/parameters/FilterQuery'
        - $ref: '#/components/parameters/SearchModeQuery'
        - $ref: '#/components/parameters/SkipQuery'
        - $ref: '#/components/parameters/LimitQuery'
      responses:
        '200':
          description: Public notes
          content:
            application/json:
              schema:
                type: object
                required: [things]
                properties:
                  things:
                    type: array
                    items:
                      $ref: '#/components/schemas/Thing'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/public/{userId}/things/{thingId}:
    get:
      tags: [public]
      summary: Get a single public note
      parameters:
        - $ref: '#/components/parameters/UserIdPath'
        - $ref: '#/components/parameters/PublicThingIdPath'
      responses:
        '200':
          description: Single public note
          content:
            application/json:
              schema:
                type: object
                required: [thing]
                properties:
                  thing:
                    $ref: '#/components/schemas/Thing'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/public/{userId}/files/{fileId}:
    get:
      tags: [public]
      summary: Legacy public file endpoint (disabled)
      description: Always returns 404 to avoid authorization bypass. Use `/api/files/{userId}/{thingId}/{identifier}` instead.
      deprecated: true
      parameters:
        - $ref: '#/components/parameters/UserIdPath'
        - name: fileId
          in: path
          required: true
          schema:
            type: string
      responses:
        '404':
          $ref: '#/components/responses/NotFound'
  /api/users:
    get:
      tags: [public]
      summary: List registered public users
      responses:
        '200':
          description: Public user list
          content:
            application/json:
              schema:
                type: object
                required: [users]
                properties:
                  users:
                    type: array
                    items:
                      $ref: '#/components/schemas/PublicUserSummary'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/users/{userId}:
    get:
      tags: [public]
      summary: Get user public profile
      parameters:
        - $ref: '#/components/parameters/UserIdPath'
      responses:
        '200':
          description: User public profile
          content:
            application/json:
              schema:
                type: object
                required: [user]
                properties:
                  user:
                    $ref: '#/components/schemas/PublicUserProfile'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/rss/{userId}:
    get:
      tags: [public]
      summary: RSS feed for public notes
      parameters:
        - $ref: '#/components/parameters/UserIdPath'
      responses:
        '200':
          description: RSS 2.0 XML feed
          content:
            application/rss+xml:
              schema:
                type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /public/{userId}:
    get:
      tags: [public]
      summary: Public HTML notes stream page
      parameters:
        - $ref: '#/components/parameters/UserIdPath'
      responses:
        '200':
          description: Rendered HTML stream page
          content:
            text/html:
              schema:
                type: string
        '400':
          $ref: '#/components/responses/BadRequest'
...